+++ lib/chat_message.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class ChatMessage extends StatelessWidget {
+  final String text;
+  final bool isUser;
+
+  const ChatMessage({super.key, required this.text, required this.isUser});
+
+  @override
+  Widget build(BuildContext context) {
+    return SelectionArea(
+      child: Container(
+        margin: const EdgeInsets.symmetric(vertical: 10.0),
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Container(
+              margin: const EdgeInsets.only(right: 16.0),
+              child: CircleAvatar(child: Text(isUser ? 'U' : 'A')),
+            ),
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    isUser ? 'ユーザー' : 'アシスタント',
+                    style: Theme.of(context).textTheme.titleSmall,
+                  ),
+                  Container(
+                    margin: const EdgeInsets.only(top: 5.0),
+                    child: Text(text),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
